quickmode fixes
